app/requirements.txt
amqp==5.1.1
anyjson==0.3.3
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==22.6.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.6
django-celery-beat==2.3.0
django-countries==7.3.2
django-filter==22.1
django-model-utils==4.2.0
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
humanize==4.2.3
isort==5.10.1
kombu==5.2.4
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
qrcode==7.3.1
redis==4.3.4
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
tomli==2.0.1
tornado==6.2
tqdm==4.64.0
typing-extensions==4.3.0
tzdata==2022.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
